packages/model-server/requirements.txt
fastapi[standard]==0.115.12
pydantic==2.11.2
sentence-transformers==4.0.2
llama-cpp-python==0.3.8